Update patch set 3
Patch Set 3:

(1 comment)

> (1 comment)
 > 
 > The oat file assistant isn't the right place for this logic. The
 > logic should be in the code that calls MakeUpToDate.

I disagree. OatFileAssistant assumes a lot of implicit locations. In particular it assumes that if you don't pass an oat_location is ok to just go and generate the oat file in dalvik_cache. Which is not the case because the permissions around that is pretty tight.

From an API perspective: I don't know where "MakeUpToDate" places files, so as a caller, I don't know if I have or not have the rights to generating anything there.

Patch-set: 3