Update patch set 2
Patch Set 2:

> I can't say I'm particularly thrilled to reduce test coverage
 > overall to get gcstress to pick it up. Is there anything else we
 > can do?

Disable this test with gcstress (As i did initially). It doesn't exercise the gc in any interesting way (simply loading a bunch of classes) and is not particularly long running otherwise.

We could also skip some of the test cases but it would be hard to ensure we still have good coverage and high repeatability if we do that (one or the other would be easy).

Patch-set: 2