Update patch set 1
Patch Set 1:

(2 comments)

I think the change must be deeper in the management of the instrumentation by the debugger. There are several methods Dbg::IsForcedInterpreter... that could take care of that. 

As discussed offline a few weeks ago, my understanding is that you want to interpreter and JIT all methods from the boot image when the debugger is connected (to avoid issues with non-debuggable optimized code).

For correctness, you should deoptimize these methods in *all* threads, not only the main thread, because you must be able to debug any thread. The best place for taht is Dbg::GoActive (which is called when the debug session is initialized). You should take a look at Instrumentation::ConfigureStubs to apply your deopt on boot image methods on startup.

Patch-set: 1