Update patch set 2
Patch Set 2:

(1 comment)

A few notes.

# Compilation time

Tested with ToT ~05/02/2016.
When enabled everywhere scheduling increases compilation time by ~5%.
When enabled only in loop blocks, the impact is negligible.


# Performance

For A57 this does not appear to affect performance overall.
Some gains and regressions can be observed in specific situations. The
regressions observed are due to execution units being oversubscribed. We should
be able to (mostly) get rid of the regressions with some additional work.

For A53 I see ~4% overall performance improvement, with peeks on maths benchmarks.
It makes sense that scheduling performs best on maths currently:
- Maths related IRs often map to few assembly instructions (often one).
- In dense maths code there are few barriers to reordering, allowing to
efficiently schedule.


# Future work

There are multiple areas we can try to work on to improve the scheduling, including:
- Schedule backwards (from last instruction in the block to first). This would allow computing the delays at the same time we visit the candidates, and offer better scheduling across barriers.
- Basic modelling of execution units to avoid oversubscribing of a unit.
- Maybe lower some IRs to avoid stall cycles within a codegen stub.
- Maybe look at dealiasing object/array accesses. (eg. accesses on a same array to different offsets don't alias).

Patch-set: 2
Label: Verified=0