Update patch set 1
Patch Set 1:

> Can you better describe the failing case? Was it kArrayObjectCheck
 > with cls.IsStackSlot()?

Right; I'll try to document this better. The regression test exhibits this issue in the case of a kArrayObjectCheck (in the first two casts), but the class to be checked against was in a register, not in a stack slot.

I'll also try to come up with a reduced test case, and also to keep using NearLabel (instead of Label) for type check kinds which are not affected by this issue.

Patch-set: 1