Update patch set 8
Patch Set 8:

OK.  I think I now understand what's going on, and you are indeed correct.
Could we add a comment along the lines of the following:

Adding 0.5 and computing the floor may fail if the value is very
near the midpoint between two integers, and the addition of 0.5 induces rounding because a larger exponent is needed for the result.
This cannot happen for negative arguments,
since the addition results in a smaller absolute value.  It cannot
happen for arguments larger than one, but smaller than the midpoint,
since the result of the addition will have the same exponent as before.
It does not matter for arguments just larger than the midpoint,
because the rounded value will still be larger than the next integer.
It does happen for the one possible value, namely one unit in the last place
less than 0.5.  And hence we handle that specially.

Patch-set: 8