Update patch set 10
Patch Set 10:

I no longer see a problem.  Thanks.

Just to complete the (moot here, but not elsewhere) Math.round discussion:

The runtime code for Math.round() was corrected in OpenJDK 8.  It's discussed in https://bugs.openjdk.java.net/browse/JDK-8010430 .

The OpenJDK 7 code that we currently have produces incorrect answers for all odd integer arguments such that the last represented digit corresponds to units.  There are a lot of those.  (This is in addition to the singleton error just below 0.5 that was corrected in OpenJDK 7 with the spec change.)

Filed b/27166445 to track the runtime problem.

Patch-set: 10