Update patch set 6
Patch Set 6: Code-Review+1

The changes to ElfBuilder are scary and make it more complicated (especially the PrepareDynamicSection method), but I do not see any obvious easier way.  I am happy with it as-is, and we can make some refactoring attempts later.

I have run my local mini-test on the ElfBuilder changes only and it seems to be producing bit-for-bit identical result as before. Yay.

Patch-set: 6
Label: Code-Review=+1