Update patch set 2
Patch Set 2:

> Any approval?

Sorry did not realize that was ready for review. Seb, are you ok with the approach?

We chatted in LON about how to approach it. And I think it's ok to just leave frames optimized on the call stack (we can't deoptimize them anyway, as HInvoke are not deoptimization points for optimizing non-debuggable).

Patch-set: 2