Update patch set 2
Patch Set 2:

I only looked at the impact of fetching the a stack map for a native pc. It looks that this CL does not impact that. You only compress the "header" of the code info while keeping the stack maps at the same size.

That seem to be good to me. I will do a full review of the CL.

To give some context: I first thought that you might introduce variable length stack maps which would increase the time needed to get to the n-th one. In the past we decided to no do it (even if it would save quite a lot of space) for time efficiency reasons.

Patch-set: 2
Label: Verified=0