Update patch set 3
Patch Set 3:

(1 comment)

The header encoding changed (and it is smaller and faster to get data from), however, that is not the main intent of the CL.  It just side effect of that I needed to change the header anyway.

The main intent is that the stack map entries are bit-packed (but still of fixed size).  So while things like native_pc, dex_pc, etc were previously using only the minimum number of bytes, they use only the minimum number of bits now.

I understand that the PC to stackmap mapping needs to be fast, and I have no intention of changing that.  Of course, I assume you know that the lookup is currently still linear because of catch blocks.  It really should be sorted and then binary searched, but that is topic for another Cl...

Patch-set: 3