Update patch set 3
Patch Set 3:

I had a chat with Calin offline. My takeaways were:

We are changing the behavior so that we will only attempt to compile code in the runtime if a DexClassLoader is used. Users who desire ahead-of-time optimized code should suggest that to the runtime by using a DexClassLoader instead of a PathClassLoader.

For the typical user, this is not a change in behavior, because we would not have had permissions to write to the global dalvik cache for the PathClassLoader case anyway. The only place this behavior could cause problems is for
1. PathClassLoader's used by processes with write permissions to the global dalvik cache. I don't think this is used anywhere? zygote? system server?
2. Users who overwrite the environment variable specifying where the global dalvik cache is, such as the ART test cases.

If either of these users desire ahead-of-time compiled code, they should switch from a PathClassLoader to a DexClassLoader to indicate that explicitly. This means we may need to change the tests, depending on what they are attempting to test.

Short term, the changes involved for the oat file assistant are:
1. Change the oat file assistant API to require an explicit oat_location for MakeUpToDate. This could be as simple as returning an error in MakeUpToDate if an explicit oat_location wasn't provided, or having a CHECK in MakeUpToDate that an explicit oat_location was provided.
2. Change the oat file manager to only call MakeUpToDate if an explicit oat_location was provided.

Long term it would be wonderful if the MakeUpToDate method was removed entirely from the OatFileAssistant, perhaps by providing a Java API for users to call when they want their code 'optimized'.

Patch-set: 3