Update patch set 1
Patch Set 1:

This is a useful feature, but I think it would be more cleanly done via the alt-interpreter.  By doing it there, you don't need to alter any individual instruction handlers so you are tracing what you would be executing if not tracing (and it works for all targets).  To make this work, you would change InitMterpTls to initialize the default handler to artMterpAsmAltInstructionStart and then add your TraceExecution() call to MterpCheckBefore.

Patch-set: 1