Update patch set 4
Patch Set 4:

Thanks for the summary Richard.

In order to DCHECK the oat_location in OatFileAssistant we do need to keep track whether or not it was initially set up (we cannot rely on the cached_oat_file_name_ since it may be updated with the default location depending on the flow). So the "simply DCHECK" does require a bit more bookkeeping. That's basically what the initial PS was doing.

So ptal and let me know what you think.

Andreas, I commented out in the run-test-jar what's expected of no-prebuild testing and added a flag to force the oat file update. Are you happy with this?

Patch-set: 4