Update patch set 3
Patch Set 3:

> > (2 comments)
 > >
 > > > Would it make more sense to add an additional argument to the
 > > > OatFileAssistant constructor indicating the target compilation
 > > > level (extract-only, full-aot, profile-guided, etc.)? You could
 > > > pass this in as an argument to getDexOptNeeded. Then the
 > existing
 > > > status result types could be used as is, interpreted with
 > respect
 > > > to the target compilation level?
 > >
 > > I am not keen to add that sort of info to the runtime to drive
 > > compilation. It adds too much complexity and exposes too many
 > > internal details which we'd have to support through our class
 > > loaders. The default assumption is that the "runtime knows the
 > best
 > > how to optimize it if needed". Compilation/extraction should
 > really
 > > happen in a separate flow.

That's an interesting idea but after some thought I have some doubts... I think this might shift the complexity in the wrong place:
- the runtime will need to decide the relation between the different forms of oat files (for example that "full-compiled" sits above "profile-compiled"). That should be a system/app strategy rather than a runtime concern (we know that different apps care about different things). 
- from an implementation perspective, we need to expose a new enum which would increase the number of combinations. Not sure if that qualifies as simpler..


 > This brings up some other cases to consider. Don't we need to have
 > refined options for PATCHOAT_NEEDED and SELF_PATCHOAT_NEEDED? It
 > may be possible to run patchoat to generate an extract-only file,
 > but not to generate a fully-compiled file, for example.

I don't think there are extra cases here. patchoat runs on the oat file, and shouldn't care how it was produced (fully or profile guide compile). There is no sense for patchoat to work on extract-only oat files since there's nothing to patch (we can dcheck that). We do plan to move everything PIC... so patchoat status here will go away. 

 > I think adding an argument to getDexOptNeeded is less complex than
 > adding more return codes whose meanings are less clear.

Which status codes do you think have a less clear meaning? We could work on improving name/documentation/granularity level so it doesn't leave place for interpretation.

 > We don't
 > need to expose this argument via class loaders, so it need not
 > expose any internal details beyond what your existing change
 > exposes to the package manager. Adding an argument to
 > getDexOptNeeded instead of changing the return status is also
 > consistent with the idea that getDexOptNeeded tells the caller how
 > to make an oat file up to date (at the target level of up to date),
 > as opposed to what the current state of the oat file is.

I wouldn't mind just re-purposing the method to return the status. The fact that we have self-patch and patch already means that the runtime is aware we cannot write to the system partition. That doesn't sound like it should be the runtime concern.

Our return is actually a status... just that before we only had 3... Is there an actual difference that you see?

 > I think if
 > you took this approach, the implementation would be simpler, with
 > less questions about what funny cases may or may not be possible.

What cases do you think are funny and raise questions? Could be that I spent too much thinking about it and things are clear to me but not to others.


Overall I don't think it makes too much of a difference between the two approaches. We are still working with private/internal api between PacakgeManager and runtime and we have flexibility. The current intent of DexOptNeeded is "Returns the VM's opinion of what kind of dexopt is needed to make the apk/jar file up to date". With the argument approach it looks like we are splitting this "opinion" across different levels.

Patch-set: 3