Update patch set 3
Patch Set 3:

> OK, I think we need to make progress with this change, and Seb and
 > I are probably thinking too far ahead on how we could clean this up
 > much further (and I'd agree with Mingyao that such cleanup does not
 > need to be addressed now).
 > 
 > So I think this change looks good to go, except
 > 1) When the debugger attaches, invalidates unconditionnally AOT
 > code of the boot image (I think Mingyao you intend to remove the
 > flag, right?)

Yes I made the change.

 > 2) When the debugger gets detached, move back boot image code to
 > AOT code.

Trying to do it and it seemed trickier than it looks. It's sort of like the uninstall case in InstallStubsForMethod() to take care of other instrumentation states, static methods of uninitialized classes, etc. It seems to me safer to just leave the interpreter bridge in place instead of going through the same complex logic to restore the oat code. Detaching debugger usually kills the process, and we have JIT anyway.

Patch-set: 3