Update patch set 1
Patch Set 1:

> Is this a stop-gap? I mean you lose testing of the result... :-)

Well, sort of. All others that test wrap-around OOB go OOB in the first few iterations. This is the only case that only goes OOB after iterating for a very long time. We still test the method with normal values, and with the checker test.

I may replace the HEAVY value with something like IsCompiled() though in the longer run.

Patch-set: 1