Update patch set 1
Patch Set 1:

We chatted about this offline.

I think it's nice to have all the compilation modes stored in the header. I also think that it makes go/ag/874183 a bit more convoluted wrt to what it decides if dex2oat is needed or not.

We decided to merge that one first and then you can update anything that's needed on top of that one.

Patch-set: 1