Update patch set 1
Patch Set 1:

There was concern previously around whether a missing store/allocation will shorten the lifetime of an object that's finalizable. Here is the rationale that it shouldn't be a concern.

Say objA is a singleton (whose allocation may be eliminated) and
objA.objB is something finalizable. And the code structure is
straight line:
    {
      <snippet>
      objA = new Wrapper();
      <snippet>
      objA.objB = new SomethingFinalizable();  // value1
      <snippet>
    }
In the above case, objA.objB is assigned in the same scope as objA
is allocated, value1 is then kept live till the end of the scope,
even if the store is eliminated. So there is no difference there
whether objA's allocation is eliminated or not.

If the code structure is (objA.objB is stored in a branch):
    {
      objA = new Wrapper();
      <snippet>
      if () {
        <snippet>
        objA.objB = new SomethingFinalizable();  // or value1
        <snippet>
      } else {
        objA.objB = value2;
      }
      <snippet N>
    }
In this case, there is a guaranteed merge of heap value of objA.objB
at <snippet N>, since value1 (a new allocation) can't be observed
other than through the first branch. objA.objB as a result won't be
eliminated (that's how the current store elimination works since we
don't create phi's for merged heap values) which in turn will make
objA's allocation necessary.

So in both cases, we should be fine.

Patch-set: 1