Update patch set 1
Patch Set 1:

Hi Chih-Hung,

is the goal of this CL to allow single pass build to use ASAN?

If yes, a single pass would still not work, because instrumented and non-instrumented binaries sometimes have common library depedencies, and non-ASAN-ified binaries should never have any ASAN-ified library deps. It means that unless all binaries are instrumented (not currently feasible), we'll always have to build these library deps twice: with and without ASAN.

If the single-pass build was not the goal, please, forgive me, I was confused by the commit message.

Patch-set: 1