Update patch set 2
Patch Set 2:

> I'm not sure that this will be faster on x86.

No worries. This is just a exploratory changelist.

Currently, round is disabled on x86 and arm since the intrinsic codegen does not have the right semantics. The idea behind this CL was to implement round() in terms of floor() and some HIR, so that we would get the implementation "for free" on all architectures. However, since floor() itself has conditional implementation on some, and none on others, not sure if we proceed with this.

In the end, we may just use the right codegen on each arch.....

Patch-set: 2
Reviewer: Aart Bik <1074526@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0