Update patch set 2
Patch Set 2:

We will use the oatdump symbolizer to make ourselves the native debug information after the fact - regardless of whether boot.oat is debuggable or not. It means people will not be able to reliably step thought the framework, but I assume that is not the primary concern of most app developers.

Patch-set: 2