Update patch set 1
Patch Set 1:

This CL is mostly for my sanity, so that we have one source of truth.
Before the CL the behavior changes depending on whether and when we attach Java debugger.
With this CL the behavior is at least consistent purely based on "am start" flags.

Based on Oleskiy's measurements from today, it appears we might need to make the compilation optional (and probably off by default) for native debugging.  Which makes it even more desirable to have one place to control it.

I am open to suggestions where the code should be placed, and how the potential system property to control the boot.oat JIT during native debugging should be called.

Patch-set: 1
Reviewer: David Srbecky <1065473@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0