Update patch set 5
Patch Set 5:

> > Hmm, don't do it like this - sets on pointers are not
 > deterministic
 > > (at least we assume so, despite arena allocation).
 > 
 > I don't follow...?
 > 
 > Re the other comment. Either way is good with me. The set should be
 > small.

ArenaSet on a pointer type is a red-black tree with nodes sorted according to pointer addresses of the entries. Running the compiler on the same piece of code multiple times could iterate over the NewInstances in a different order each time. That shouldn't happen given the implementation of ArenaAllocator but we try to avoid doing this nonetheless for debuggability. Typically we'd use a set on IDs instead of a set on pointers because those don't change between runs, but that's not an option here.

Patch-set: 5
Label: Code-Review=0