Update patch set 12
Patch Set 12:

> lgtm if you remove the SsaBuilder dependency from the
 > InstructionBuilder as we discussed.

Thanks Nicolas. I looked more into it and the dependency is not just on the ambiguous/uninitialized lists but also on GetFloat/ReferenceEquivalentOf. That's a typing-relevant code that I want to keep near the rest of the primitive type propagation. If I'd want to break that dependency, I'd therefore have to get rid of SsaBuilder entirely and move all of its code into HGraphBuilder.

Patch-set: 12
Label: Verified=0