Update patch set 1
Patch Set 1:

> > > I don't think this is a good change, as it makes everything
 > > > messier. But that's up to you.
 > >
 > > Don't you agree we're currently paying a non negligible cost by
 > > going through the instrumentation framework?
 > 
 > I don't think I agree.
 > 
 > Considering that dispatch is expensive anyways (setting up a shadow
 > frame etc), I don't think going through a single-element vector and
 > dispatching is significant.

You have to agree, it does add up...

 > 
 > mterp is already special-cased to not do that for all internal
 > steps.

Indeed, so if you take it from the JIT POV, I think it's pretty unclean for it to have two ways to collect samples: one directly through mterp, one indirectly with the other interpreters.

 > 
 > I'd rather keep the other interpreters and the driver code as clean
 > as possible (especially with the plan to move "fast" dispatch into
 > mterp, which can then, again, special-case).

I'm sympathetic to this preference. Main problem I see is that we already need a special cases even for the other interpreters for on stack replacement. OSR doesn't nicely fit in the instrumentation framework. I also plan on giving weight to compiled/interpreted transitions, so once again the sampling will be hard coded.

 > 
 > (But as mentioned, I'm not gonna stand in the way of this, just
 > wanted to voice my general opposition...)

That is fine, helps me explain more why I'd like to go in the opposite direction.

Patch-set: 1
Label: Verified=0