Update patch set 6
Patch Set 6:

(1 comment)

PTAL, I used Calin/Nicolas' idea of making sure that trying to create in erroneous type from anywhere yields the inexact object type instead. This required a new interceptor method, but I think it was worth it, as it seems this also exposed a few lock errors.

With unit tests, of course!

Patch-set: 6