Update patch set 7
Patch Set 7:

> > Ok, I hit another hurdle with this approach :-(
 > >
 > >
 > > For the scanner case under consideration, this crashed while
 > trying
 > > to set
 > >
 > > SetLoadedClassRTI()
 > >
 > > since this DCHECKs on exact. I see a few more checks on this on
 > > related methods on LoadClass.
 > >
 > > Any pointers on what to do there?
 > 
 > Ah, that means that the class we are trying to load is errornous.
 > That's weird but maybe I'm missing something about errornous
 > classes.
 > I'll have to think when that can happen.

I thought more about this and it doesn't look like this is going in the right direction.

Fixing IsObjectClass is what we should care about. Assuming that IsObjectClass and GetCommonSuperClass is correct, RTP should be fine even in the presence of errornous types. We may go one step beyond and actually transform every errornous type to inexact object (as you are trying to do here) but that's not really needed because it invalidates sensible invariants that we added.

What do you think?

Patch-set: 7