Update patch set 4
Patch Set 4:

The floor(a + 0.5f) code is wrong.  Don't go there.  It also fails for many large integers for which a + 0.5f rounds to a + 1.

See the comment by Alexey (Mar 7) together with my comment 3 after that (Mar 9) for a correct algorithm.

There was some debate in the context of another architecture as to whether that's worth implementing.

Patch-set: 4
Label: Presubmit-Verified=0