Update patch set 4
Patch Set 4:

Hans, is the link https://buganizer.corp.google.com/issues/26327751 supposed to be readable by the public. I'm logged in so I can access this web site but when I go to this URL I'm presented with a login screen. I'd like to be able to see the discussion about on this. Does this URL require a different set of credentials?

Patch-set: 4