Update patch set 2
Patch Set 2:

Could you add a test with this change. The test can:
1) create a thread and start it
2) on the created thread enter JNI code
3) in the JNI code lock an object
4) return from the JNI code and terminate the thread
5) on the main thread do a synchronized on the object the JNI code did the monitor enter upon
I suspect the issue is that a detaching thread isn't having its monitors exited as the monitor was never inflated. You should be able to add logging to confirm this from the problem in the bug.
As monitors should be scoped (checked by the method verifier) or enter/exited in JNI, having an object locked by a deceased thread is evidence of a bug elsewhere in the runtime and a work around in the thread suspension unnecessary although a source of extra robustness.
It would also be possible to introduce extra diagnostics for this situation when check JNI is enabled.

Patch-set: 2