Update patch set 3
Patch Set 3:

> > I'd say that's something for dex version 37, and a clarification
 > of
 > > the spec.
 > 
 > What do you want to apply from version 37 onward? Just the
 > CheckSizeLimit() or also the proto ordering changes?
 > 
 > Do you want to explicitly mention the lexicographical ordering of
 > argument type_list in the documentation?

The check-size-limit, as right now we would accept a dex file with more type-ids (that are just unused). I thought the proto ordering changes somewhat rely on this.

Maybe request a market scan. If it's (expectedly) clean, we could do it for v35, in general.

Patch-set: 3
Label: Verified=0