Update patch set 1
Patch Set 1:

Good point.  I had missed that one.

This one tests a bit more of the stack and, by forcing more expansion, may be a slightly better test of the id <-> address mapping code.

This runs quickly.  Is there a down side to adding it, too?

Patch-set: 1
Label: Verified=0