Update patch set 3
Patch Set 3:

> (1 comment)
 > 
 > Note that I don't think we want to proceed with this change, at
 > least for fixing b/27831001. Maybe we can just keep the ArtField in
 > the FieldInfo from that change.

I'm a little confused on what is "that change"? This change is the one that adds ArtField into FieldInfo, and that fixes the gvn part for b/27831001 (plus the additional comparison of the ArtField in InstructionDataEquals()). There is a different fix for the gvn part now?

Patch-set: 3
Reviewer: Mingyao Yang <1043514@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0