Update patch set 3
Patch Set 3:

Hi Nicolas, as I understand the current problem this CL fixes is described by DCHECK Yevgeny added at the last patchset in code generator. Specifically it might happen that instruction in the method has an environment of different size than others. So during generation of stack maps in oat file they (stack maps) will be of different size and that results in decoding of the stack maps will be broken at all. Yevgeny, is that correct explanation?

Patch-set: 3