Update patch set 2
Patch Set 2:

> Try to chose a more selective group of values, or
 > otherwise craft the tests with a loop over an interesting range of
 > values.

We have a bit of a problem here. We either have to write a manageable amount (several dozens tops) of tests with hard-coded inputs/outputs or we have to programmatically validate the output if inputs are generated in a loop or otherwise. The latter implies either duplicating the logic of LoadConst64() in the test (which is kind of pointless) or writing a simple MIPS64 emulator for the handful of instructions that LoadConst64() emits. If you're OK with this latter approach, we can do it. Yet another alternative is to test LoadConst64() on the target, but I think it may be more hassle to generate assembly or Java test code, exercising LoadConst64(). Or not. What do you think?

Patch-set: 2