Update patch set 6
Patch Set 6: Code-Review+1

A test is added.
I found that the bug cannot cause the incorrect StackMaps generated because the inliner is switched off for --debuggable but only for --debuggable additional StackMaps are generated for all instructions that have environment. In other words, the bug is not critical.
This will be changed with the patch I1afaf275c62d2f72fc3aea22852e87d830dc237c that generates debug info for non-debuggables.

Patch-set: 6
Label: Code-Review=+1