Update patch set 1
Patch Set 1:

Unsafe stack overflow is only tested in the interpreter mode with this change. In the "optimizing" mode the test still passes, but the actual overflow happens on the system stack.

Maybe someone could help me write a better test? Basically, we need some java code (not JNI!) that uses the unsafe stack. This means that the code (when running without the interpreter) must call something from runtime (libart or anything built with the system compiler) and the runtime should call back into Java.

Patch-set: 1